Move status endpoint test to a dedicated file
To improve the organization of tests, the status endpoint
test is moved from `usearch.rs` to a new `status.rs` file.

This keeps test files focused on a single responsibility and makes it
easier to add related tests in the future.
